Shim implementation
This commit contains the shim implementation, which includes a function
for the consumer to create an os/exec.(*Cmd) from a run specification to
launch the shim with the appropriate configuration, together with a simple
reexec implementation that recognizes when the shim has been launched and
redirects control to the shim's entry point.

Tests ensure the shim launches a process in precisely the namespaces
requested in the requirements doc, and shim stopping functionality.