Server implementation
This commit contains the implementation of the WorkerServiceServer gRPC
service. The Server type also doubles as the API server, setting up a
listener socket and starting the gRPC server.

Tests include a variety of authn/authz scenarios as well as a sanity
test that ensures the server is able to return the stdout and stderr
of a process to the client.